# FreeTimeApp/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "tiempo-libre",
    "version": "1.0"
  },
  "paths": {
...
    "/api/bloques-reservacion/generar": {
      "post": {
        "tags": [
          "BloquesReservacion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneracionBloquesRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneracionBloquesRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GeneracionBloquesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/bloques-reservacion/aprobar": {
      "post": {
        "tags": [
          "BloquesReservacion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AprobarBloquesRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AprobarBloquesRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AprobarBloquesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/bloques-reservacion/cambiar-empleado": {
      "post": {
        "tags": [
          "BloquesReservacion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarBloqueRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarBloqueRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarBloqueRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/bloques-reservacion": {
      "get": {
        "tags": [
          "BloquesReservacion"
        ],
        "parameters": [
          {
            "name": "anioObjetivo",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "grupoId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "areaId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "estado",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "soloConEspacio",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/bloques-reservacion/empleado/{empleadoId}": {
      "get": {
        "tags": [
          "BloquesReservacion"
        ],
        "parameters": [
          {
            "name": "empleadoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "anioObjetivo",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/bloques-reservacion/estadisticas": {
      "get": {
        "tags": [
          "BloquesReservacion"
        ],
        "parameters": [
          {
            "name": "anioObjetivo",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "grupoId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "areaId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/bloques-reservacion/por-fecha": {
      "get": {
        "tags": [
          "BloquesReservacion"
        ],
        "parameters": [
          {
            "name": "fecha",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "areaId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "grupoId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "anioObjetivo",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/bloques-reservacion/eliminar": {
      "delete": {
        "tags": [
          "BloquesReservacion"
        ],
        "parameters": [
          {
            "name": "anioObjetivo",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "grupoIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/bloques-reservacion/empleados-no-respondieron": {
      "get": {
        "tags": [
          "BloquesReservacion"
        ],
        "parameters": [
          {
            "name": "anioObjetivo",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "areaId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "grupoId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AprobarBloquesRequest": {
        "required": [
          "anioObjetivo",
          "grupoIds"
        ],
        "type": "object",
        "properties": {
          "anioObjetivo": {
            "type": "integer",
            "format": "int32"
          },
          "grupoIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "observaciones": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CambiarBloqueRequest": {
        "required": [
          "bloqueDestinoId",
          "bloqueOrigenId",
          "empleadoId",
          "motivo"
        ],
        "type": "object",
        "properties": {
          "empleadoId": {
            "type": "integer",
            "format": "int32"
          },
          "bloqueOrigenId": {
            "type": "integer",
            "format": "int32"
          },
          "bloqueDestinoId": {
            "type": "integer",
            "format": "int32"
          },
          "motivo": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "observacionesAdicionales": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GeneracionBloquesRequest": {
        "required": [
          "anioObjetivo",
          "fechaInicioGeneracion"
        ],
        "type": "object",
        "properties": {
          "fechaInicioGeneracion": {
            "type": "string",
            "format": "date-time"
          },
          "anioObjetivo": {
            "maximum": 2030,
            "minimum": 2024,
            "type": "integer",
            "format": "int32"
          },
          "grupoIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "soloSimulacion": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}